Fix compiler warning

+++ win/curses/cursinvt.c
@@ -82,7 +82,7 @@ curses_add_inv(int y, int glyph, CHAR_P accelerator, attr_t attr,
     if (accelerator && glyph != NO_GLYPH && iflags.use_menu_glyphs) {
         unsigned dummy = 0; /* Not used */
         int color = 0;
-        int symbol = 0;
+        long int symbol = 0;
         mapglyph(glyph, &symbol, &color, &dummy,
                      u.ux, u.uy);
         attr_t glyphclr = curses_color_attr(color, 0);
